DNSimple: TXT length and derived-record handling
The new length limit for TXT records is documented as being 1000 "characters"
(and doesn't specify how characters are counted).

Drop a comment on the generated TXT records suggesting a possible better
approach for filtering them; if a future code editor sees another scenario
where implicitly generated records are showing up, this might be the clue they
need to fix the provider.  But since DNSimple don't document that API field
that I can find, leave the actual logic on what's known to work.